added small refactor changes
